Bug 1277653 - Add support for filtering whether tests should be run based upon which directories were changed in a build; r=bc
This adds a "run_if_changed" item to runtests section of a test configuration
file. If present, the test will only run if files under those paths were
modified in the revision being tested.

This requires getting the pushlog and the diffs from the mercurial server in
order to construct a list of files that were changed by that revision (and any
other revision pushed at the same time.)

If we are unable to retrieve the pushlog or the list of files, the test will
run anyway, under the assumption that is both safer and a more likely to be
noticed error condition than skipping the tests.