Bug 1349904 - Autophone - Do not consider missing minidumps directory as a fatal error, r=jmaher.
We have been treating a missing minidumps directory in the profile as an error even
immediately after the profile was created and initialized.

This has resulted in a variety of oranges particularly with "Failure initializing profile"
with no real action plan on if this is truly an error or something that can be fixed.

This patch removes the error but continues to log the issue whenever
AutophoneCrashProcessor.handle_crashes() is called so that we might
continue to track the issue.